# api/openapi.yaml
openapi: 3.0.3
info:
  title: BS/AD Calendar API
  version: 1.0.0
  description: |
    API-controlled Bikram Sambat (BS) and Gregorian (AD) calendar: year data, events and
    admin-controlled UI configuration for web (Next.js) and mobile (React Native) clients.

    This document is the **contract**. The server embeds it, serves it at `/openapi.yaml`,
    and the flow test (`services/calendar-api/internal/httpapi/flow_test.go`) validates every
    real request and response against it, so the docs cannot drift from the implementation.

    ## How clients use the API

    1. Ship the bundled year table and default theme, and render immediately.
    2. Call `GET /v1/manifest` on start, on foreground and every 15 minutes (it is tiny and CDN-cached for 30 s).
    3. Download only what changed. Every other URL contains its version and is **immutable**:
       `/v1/calendar/data/{version}`, `/v1/ui-config/{app}/{version}`, `/v1/events/buckets/{bsYear}/{version}`.
    4. Convert dates locally with the year table. `GET /v1/convert` exists for thin clients and other backends.

    See `docs/api.md` for the full guide, including curl walkthroughs of every flow.

    ## Authentication

    * **Public routes** need an API key in `X-Api-Key`. Public keys (`pk_…`) ship inside apps and are not
      secrets; they identify the app, enforce origin allow-lists and rate limits. Server keys (`sk_…`) are for backends.
...

    ## Conventions

...
      never typed as `format: date`.
...
  contact:
...
servers:
...
paths:
...
  /v1/manifest:
    get:
      tags: [Sync]
      summary: Current versions of everything a client caches
      description: |
        The only frequently-changing resource. Compare each version with your cache and fetch only
        what changed. Cached at the CDN for 30 seconds; send `If-None-Match` to get `304`.

        Bucket version for a BS year = `eventBuckets[year]` if present, otherwise `defaultBucketVersion`.
        A config `candidate` is a staged rollout: apply it only if `hash(installId) % 100 < percent`.
      operationId: getManifest
      parameters:
        - name: app
          in: query
          description: App key whose UI config to report (for example `web` or `mobile`).
          schema: { type: string, pattern: '^[a-z][a-z0-9-]{1,30}$', default: web }
        - name: clientVersion
          in: query
          description: Semver of the calling app. Enables `updateRequired` and config compatibility.
          schema: { type: string, pattern: '^\d+\.\d+\.\d+$', example: 1.4.0 }
      responses:
        '200':
          description: Manifest.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
            Cache-Control: { $ref: '#/components/headers/CacheControl' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Manifest' }
        '304': { $ref: '#/components/responses/NotModified' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        default: { $ref: '#/components/responses/Problem' }
  /v1/calendar/data/latest:
    get:
      tags: [Sync]
      summary: Redirect to the current year table
      operationId: getCalendarDataLatest
      responses:
        '302':
          description: Redirect to `/v1/calendar/data/{currentVersion}` (not cached).
          headers:
            Location: { schema: { type: string } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Problem' }
  /v1/calendar/data/{version}:
    get:
      tags: [Sync]
      summary: Year table (immutable)
      description: |
        The full BS year table for a data version. Immutable: cache forever.
        Verify `sha256` (see the CalendarData schema for the exact canonical form) and the table invariants
        before using it; on failure keep the previous table.
      operationId: getCalendarData
      parameters:
        - name: version
          in: path
          required: true
          schema: { type: integer, format: int64, minimum: 1 }
      responses:
        '200':
          description: Year table.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
            Cache-Control: { $ref: '#/components/headers/CacheControl' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CalendarData' }
        '304': { $ref: '#/components/responses/NotModified' }
        '404': { $ref: '#/components/responses/NotFound' }
        default: { $ref: '#/components/responses/Problem' }
  /v1/ui-config/{app}/{version}:
    get:
      tags: [Sync]
      summary: UI configuration (immutable)
      description: Version `0` is the built-in default and is cached for an hour; published versions are immutable.
      operationId: getUiConfig
      parameters:
        - $ref: '#/components/parameters/App'
        - name: version
          in: path
          required: true
          schema: { type: integer, minimum: 0 }
      responses:
        '200':
          description: The configuration document (see `fixtures/ui-config.schema.json`).
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
            Cache-Control: { $ref: '#/components/headers/CacheControl' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UiConfigDocument' }
        '304': { $ref: '#/components/responses/NotModified' }
        '404': { $ref: '#/components/responses/NotFound' }
        default: { $ref: '#/components/responses/Problem' }
  /v1/events/buckets/{bsYear}/{version}:
    get:
      tags: [Sync]
      summary: All published events touching a BS year (immutable)
      description: |
        Recurrences are expanded. An AD or BS month view needs at most two buckets.
        Requesting an older version redirects (`302`, not cached) to the current one;
        a future version returns `404 VERSION_NOT_FOUND`.
      operationId: getEventBucket
      parameters:
        - name: bsYear
          in: path
          required: true
          schema: { type: integer, example: 2083 }
        - name: version
          in: path
          required: true
          schema: { type: integer, format: int64, minimum: 0 }
      responses:
        '200':
          description: Bucket.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
            Cache-Control: { $ref: '#/components/headers/CacheControl' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EventBucket' }
        '302':
          description: The version is outdated; follow `Location` to the current one.
          headers:
            Location: { schema: { type: string } }
        '304': { $ref: '#/components/responses/NotModified' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        default: { $ref: '#/components/responses/Problem' }

  # ---------------------------------------------------------------- Dates
  /v1/convert:
    get:
      tags: [Dates]
      summary: Convert one date between AD and BS
      description: Send exactly one of `ad` or `bs`. Devanagari digits are accepted.
      operationId: convertDate
      parameters:
        - name: ad
          in: query
          schema: { type: string, example: '2026-09-24' }
        - name: bs
          in: query
          schema: { type: string, example: '2083-06-08' }
      responses:
        '200':
          description: The date in both calendars.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Conversion' }
              example:
                ad: { date: '2026-09-24', year: 2026, month: 9, day: 24, monthName: { en: September, ne: सेप्टेम्बर } }
                bs: { date: '2083-06-08', dateNe: २०८३-०६-०८, year: 2083, month: 6, day: 8, monthName: { en: Ashwin, ne: असोज }, daysInMonth: 31, yearStatus: verified }
                weekday: { index: 4, name: { en: Thursday, ne: बिहीबार } }
                epochDay: 20720
                dataVersion: 1
        '304': { $ref: '#/components/responses/NotModified' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        default: { $ref: '#/components/responses/Problem' }
  /v1/today:
    get:
      tags: [Dates]
      summary: Today in both calendars
      operationId: getToday
      parameters:
        - name: tz
          in: query
          description: IANA time zone that defines "today".
          schema: { type: string, default: Asia/Kathmandu }
      responses:
        '200':
          description: Today.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Conversion' }
        '304': { $ref: '#/components/responses/NotModified' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        default: { $ref: '#/components/responses/Problem' }
  /v1/months/{basis}/{year}/{month}:
    get:
      tags: [Dates]
      summary: Ready-to-render 42-cell month grid
      operationId: getMonthGrid
      parameters:
        - name: basis
          in: path
          required: true
          schema: { type: string, enum: [AD, BS] }
        - name: year
          in: path
          required: true
          schema: { type: integer }
        - name: month
          in: path
          required: true
          schema: { type: integer, minimum: 1, maximum: 12 }
        - name: weekStart
          in: query
          description: 0 = Sunday (Nepal default), 1 = Monday, 6 = Saturday.
          schema: { type: integer, minimum: 0, maximum: 6, default: 0 }
        - name: weekendDays
          in: query
          description: Comma-separated weekdays (0-6). Nepal default is Saturday only.
          schema: { type: string, default: '6' }
        - name: include
          in: query
          description: Set to `events` to attach published events to each cell.
          schema: { type: string, enum: [events] }
      responses:
        '200':
          description: Month grid.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MonthGrid' }
        '304': { $ref: '#/components/responses/NotModified' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        default: { $ref: '#/components/responses/Problem' }

  # ---------------------------------------------------------------- Events
  /v1/events:
    get:
      tags: [Events]
      summary: Published events in a date range (max 366 days)
      description: For servers and search. Apps should use buckets from the manifest instead.
      operationId: listPublicEvents
      parameters:
        - name: from
          in: query
          required: true
          schema: { type: string, example: '2083-06-01' }
        - name: to
          in: query
          required: true
          schema: { type: string, example: '2083-06-31' }
        - name: basis
          in: query
          description: Calendar of `from` and `to`.
          schema: { type: string, enum: [AD, BS], default: AD }
        - name: category
          in: query
          description: Comma-separated category keys.
          schema: { type: string }
        - name: q
          in: query
          description: Case-insensitive title search (English or Nepali).
          schema: { type: string }
      responses:
        '200':
          description: Matching occurrences, sorted by start date.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EventRange' }
        '304': { $ref: '#/components/responses/NotModified' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        default: { $ref: '#/components/responses/Problem' }
  /v1/events.ics:
    get:
      tags: [Events]
      summary: iCalendar subscription feed
      description: |
        Subscribe from Google Calendar, Apple Calendar or Outlook. Calendar apps cannot send headers,
        so this route also accepts the key as `?api_key=`. Use a public key restricted to this purpose.
      operationId: getICS
      security:
        - ApiKey: []
        - ApiKeyQuery: []
      parameters:
        - name: fromYear
          in: query
          description: First BS year (default current year - 1).
          schema: { type: integer }
        - name: toYear
          in: query
          description: Last BS year (default current year + 1, at most fromYear + 4).
          schema: { type: integer }
        - name: category
          in: query
          schema: { type: string }
        - name: lang
          in: query
          schema: { type: string, enum: [en, ne], default: en }
      responses:
        '200':
          description: RFC 5545 calendar.
          content:
            text/calendar:
              schema: { type: string }
        '304': { $ref: '#/components/responses/NotModified' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        default: { $ref: '#/components/responses/Problem' }
  /v1/categories:
    get:
      tags: [Events]
      summary: Event categories (legend colours and holiday flags)
      operationId: listPublicCategories
      responses:
        '200':
          description: Categories in display order.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [categories]
                properties:
                  categories:
                    type: array
                    items: { $ref: '#/components/schemas/PublicCategory' }
        '304': { $ref: '#/components/responses/NotModified' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Problem' }

  # ---------------------------------------------------------------- Telemetry
  /v1/telemetry:
    post:
      tags: [Telemetry]
      summary: Report client health events
      description: Anonymous, batched (1-50 events). Feeds the `client_telemetry_events_total` metric and alerts.
      operationId: postTelemetry
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TelemetryBatch' }
      responses:
        '202':
          description: Accepted.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [accepted]
                properties:
                  accepted: { type: integer }
        '422': { $ref: '#/components/responses/Unprocessable' }
        default: { $ref: '#/components/responses/Problem' }

  # ---------------------------------------------------------------- Admin auth
  /v1/admin/auth/login:
    post:
      tags: [Admin auth]
      summary: Sign in with email and password
      description: Rate-limited to 10 attempts per minute per IP. Wrong email and wrong password look identical.
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email, password]
              properties:
                email: { type: string, format: email, example: admin@example.com }
                password: { type: string, format: password }
      responses:
        '200':
          description: Tokens. Store the refresh token server-side (for example in an httpOnly cookie of the admin panel's backend).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        default: { $ref: '#/components/responses/Problem' }
  /v1/admin/auth/refresh:
    post:
      tags: [Admin auth]
      summary: Rotate the refresh token and get a new access token
      description: Each refresh token works once. Presenting an already-used token revokes the session (`TOKEN_REUSED`).
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200':
          description: New tokens.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Problem' }
  /v1/admin/auth/logout:
    post:
      tags: [Admin auth]
      summary: Revoke the current session
      operationId: logout
      security: [{ BearerAuth: [] }]
      responses:
        '204': { description: Signed out. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Problem' }
  /v1/admin/me:
    get:
      tags: [Admin auth]
      summary: The signed-in user and their permissions
      operationId: getMe
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Current user.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Me' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Problem' }

  # ---------------------------------------------------------------- Admin categories & events
  /v1/admin/categories:
    get:
      tags: [Admin events]
      summary: List categories
      operationId: listCategories
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Categories.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/Category' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Problem' }
    post:
      tags: [Admin events]
      summary: Create a category
      description: Requires `categories:write`. Category changes bump every bucket version (colours appear in buckets).
      operationId: createCategory
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CategoryInput' }
      responses:
        '201':
          description: Created.
          headers:
            Location: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Category' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        default: { $ref: '#/components/responses/Problem' }
  /v1/admin/categories/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    patch:
      tags: [Admin events]
      summary: Update a category (JSON Merge Patch)
      description: The `key` cannot change because clients filter by it.
      operationId: patchCategory
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema: { $ref: '#/components/schemas/CategoryPatch' }
          application/json:
            schema: { $ref: '#/components/schemas/CategoryPatch' }
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Category' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        default: { $ref: '#/components/responses/Problem' }
    delete:
      tags: [Admin events]
      summary: Delete an unused category
      operationId: deleteCategory
      security: [{ BearerAuth: [] }]
      responses:
        '204': { description: Deleted. }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        default: { $ref: '#/components/responses/Problem' }
...
  /v1/admin/events/import:
    post:
      tags: [Admin events]
      summary: Bulk import from CSV (dry run by default)
      description: |
        Columns: `category,title_en,title_ne,basis,start,end,all_day,start_time,end_time,recurrence,rrule,recur_until,is_holiday,description_en,description_ne`
        (`category`, `title_en`, `basis`, `start` required). Every row is validated first; with `dryRun=false`
        all valid rows are created as drafts in one transaction, and nothing is created if any row is invalid.
      operationId: importEvents
      security: [{ BearerAuth: [] }]
      parameters:
        - name: dryRun
          in: query
          schema: { type: boolean, default: true }
        - name: skipDuplicates
          in: query
          schema: { type: boolean, default: false }
      requestBody:
        required: true
        content:
          text/csv:
            schema: { type: string }
      responses:
        '200':
          description: Dry-run report.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ImportReport' }
        '201':
          description: Imported.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ImportReport' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        default: { $ref: '#/components/responses/Problem' }
  /v1/admin/events/copy-year:
    post:
      tags: [Admin events]
      summary: Copy one-off BS events to another year as drafts
      description: Lunar festivals move every year, so review each draft before publishing. Safe to run twice.
      operationId: copyYear
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CopyYearInput' }
      responses:
        '201':
          description: Report.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CopyYearReport' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        default: { $ref: '#/components/responses/Problem' }
...
  /v1/admin/years:
    get:
      tags: [Admin year table]
      summary: The current year table with status and sources
      operationId: listYears
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Years.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/YearList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Problem' }
...
security:
  - ApiKey: []
components:
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-Api-Key
      description: Public (`pk_…`) or server (`sk_…`) API key. Local dev key `pk_dev_local_public_key_0001`.
    ApiKeyQuery:
      type: apiKey
      in: query
      name: api_key
      description: Accepted only on `/v1/events.ics`.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...